requirements.txt
APScheduler==3.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==0.12
googletrans==2.4.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
lxml==3.7.2
MarkupSafe==1.1.0
pahk-Win32w==0.1.0
pyperclip==1.7.0
pypinyin==0.34.0
pytz==2018.7
requests==2.20.0
six==1.12.0
tzlocal==1.5.1
urllib3==1.24.1
Werkzeug==0.11.15


pymongo
redis


